Wemo plugin support (#3184)
* Add wemo plugin

Similar to the other PSU control plugins, this adds support for the
wemoswitch plugin (https://github.com/jneilliii/OctoPrint-WemoSwitch).
The plugin requires the IP and port of the wemo switch you want to
control and works exactly the same as the other PSU control plugins.

* Update locale information for wemo plugin

I kept this as a separate commit since it seems to change more than just
my additions so it's easier to revert if necessary

* Formatting; allow installing Wemo during install script

* Lint

---------

Co-authored-by: Queen Vinyl Da.i'gyu-Kazotetsu <vinyldarkscratch@gmail.com>